+++ frontend/src/hooks/useFetchTeam.ts
@@ -1,0 +1,36 @@
+import { useState, useEffect } from "react";
+import { useParams } from "react-router-dom";
+import Cookies from "js-cookie";
+import { AxiosRequestConfig, AxiosResponse } from "axios";
+import { axiosInstance } from "../utils/axios";
+import { Team, TeamsShowResponse, User } from "../types";
+
+export const useFetchTeam = () => {
+  const params = useParams<{ id: string }>();
+  const [team, setTeam] = useState<Team>();
+  const [users, setUsers] = useState<User[]>();
+
+  const options: AxiosRequestConfig = {
+    url: `/teams/${params.id}`,
+    method: "GET",
+    headers: {
+      "access-token": Cookies.get("_access_token") || "",
+      client: Cookies.get("_client") || "",
+      uid: Cookies.get("_uid") || "",
+    },
+  };
+
+  useEffect(() => {
+    axiosInstance(options)
+      .then((res: AxiosResponse<TeamsShowResponse>) => {
+        console.log(res);
+        setTeam(res.data.team);
+        setUsers(res.data.users);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  }, []);
+
+  return { team, users };
+};
